requirements.txt
Pillow==9.2.0
pickle5==0.0.11
streamlit==1.22.0
numpy==1.24.3
pandas==1.5.1
altair<5